src/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
chromadb==0.5.15
sentence-transformers==3.1.1
langchain==0.3.7
langchain-community==0.3.7
openai==1.51.2
scikit-learn==1.5.2
numpy==1.26.4
requests==2.32.3
